CLI-1276: no-scripts doesn't disable composer hooks
